refactor(dev): move server plugins to serverjs namespace
